refactor: move FuelTank, ChemicalTank, ElectricTank into fueltank.py
